Moving the cheatsheets into separate dirs
