+++ examples/src/main/kotlin/cheatsheets/basics/1_Main.kt
@@ -1,4 +1,4 @@
-package cheatsheets
+package cheatsheets.basics
 
 /**
  * Themen:
