+++ examples/src/main/kotlin/cheatsheets/basics/5_Methods.kt
@@ -1,4 +1,4 @@
-package cheatsheets
+package cheatsheets.basics
 
 /**
  * Themen:
