+++ examples/src/main/kotlin/cheatsheets/basics/6_oop_simple.kt
@@ -1,4 +1,4 @@
-package cheatsheets
+package cheatsheets.basics
 
 /**
  * Themen:
