+++ examples/src/main/kotlin/cheatsheets/basics/7_fp_Lambdas.kt
@@ -1,4 +1,4 @@
-package cheatsheets
+package cheatsheets.basics
 
 /**
  * Themen:
