+++ examples/src/main/kotlin/cheatsheets/oop/10_oop_inheritance.kt
@@ -1,5 +1,4 @@
-package cheatsheets
-
+package cheatsheets.oop
 /**
  * Eine "normale" Klasse haben wir bereits kennen gerlent:
  */
