+++ examples/src/main/kotlin/cheatsheets/oop/11_oop_customAccessors.kt
@@ -1,4 +1,4 @@
-package cheatsheets
+package cheatsheets.oop
 
 class Car {
